goConcurrency: Add Makefile ( -race flag)
go run -race main.go
-race flag helps detect any race condition due to the code.

"When the -race command-line flag is set, the compiler instruments
all memory accesses with code that records when and how the memory
was accessed, while the runtime library watches for unsynchronized
accesses to shared variables. When such "racy" behavior is detected,
a warning is printed. (See this article for the details of the algorithm.)"

https://blog.golang.org/race-detector