OOG: constructors in go
https://stackoverflow.com/questions/18125625/constructors-in-go
https://golangbot.com/structs-instead-of-classes/

Go doesn't support constructors. If the zero value of a type is
not usable, it is the job of the programmer to unexport the type
to prevent access from other packages and also to provide a
function named NewT(parameters) which initialises the type T with
the required values.

Provide a function func New() YourTyp or if you have several such
types in your package functions func NewYourType1() YourType1 and
so on.